make giflib/intrafont a submodule and move special
