+++ src/image/format/oslWriteImageFilePNG.c
@@ -1,9 +1,9 @@
-#include "../oslib.h"
-#include "../png.h"
-#include "../pngconf.h"
-#include "../zlibInterface.h"
-#include "../zlib.h"
-#include "../zconf.h"
+#include "../../oslib.h"
+#include "../../png.h"
+#include "../../pngconf.h"
+#include "../../zlibInterface.h"
+#include "../../zlib.h"
+#include "../../zconf.h"
 
 void oslPngFlushFn(png_structp png_ptr);
