+++ src/image/oslLoadImageFile.c
