+++ src/image/oslWriteImageFile.c
